web app repository: findAll only with pagination

+++ vlo-5.0/vlo-web-app/src/main/java/eu/clarin/cmdi/vlo/web/service/VloApiClient.java
@@ -26,6 +26,6 @@ import reactor.core.publisher.Flux;
  */
 public interface VloApiClient {
 
-    Flux<VloRecord> getRecords(String q, Optional<Integer> rows, Optional<Integer> start);
+    Flux<VloRecord> getRecords(String q, Optional<Long> rows, Optional<Long> start);
 
 }
